feat: fix some bugs

+++ database/migrations/1684845133745_boxes.ts
@@ -8,7 +8,7 @@ export default class extends BaseSchema {
       table.increments('id')
 
       table.string('uid').notNullable()
-      table.bigInteger('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
+      table.integer('user_id').unsigned().references('id').inTable('users').onDelete('CASCADE')
 
       table.string('name').notNullable()
